Use wait_for_connection for the Heat nodes
The `wait_for_connection` module is more reliable as it uses Ansible's `ping`
to verify the nodes are really accessible. Using `wait_for` and checking that
port 22 is open runs into the possibility of SSH being up but the public keys
or users not being set up yet (as that's done with cloud-init).

In addition, we were gathering facts before running the wait_for task which
rendered it useless.