Set up NetworkManager automatically (#542)
* Set up NetworkManager automatically

This removes the extra step of running the
`openshift-ansible/playbooks/byo/openshift-node/network_manager.yml`
before installing openshift. In addition, the playbook relies on a
host group that the provisioning doesn't provide (oo_all_hosts).

Instead, we set up NetworkManager on CentOS nodes automatically. And
we restart it on RHEL (which is necessary for the nodes to pick up the
new DNS we configured the subnet with).

This makes the provisioning easier and more resilient.

* Apply the node-network-manager role to every node

It makes the code simpler and more consistent across distros.