Remove clouds.yaml from sample-inventory
With the move to the static inventory, we don't need it anymore so it's
now just an unnecessary step in the deployment.

Note that the users may still want to use clouds.yaml for openstack
credentials instead of sourcing the `OS_*` environment variables, but
they can do that at their discression.

The reason we had the clouds.yaml here was because the `openstack.py`
dynamic inventory used the servers' UUID's as ansible hosts by default
and the options we put in caused it to use the hostnames (as desired).