Only fetch the master branch once
We used to load the master branch every time we sourced vars.sh, but
that file should not have any side effects.

Also, since the run takes up to an hour to finish, this meant the output
could differ if something merged in the meantime.

So with this we only fetch the master and print the modified files once
and vars.sh just reads the status of the repo afterwards.