Set Ansible version in openstack CI for 2.3 (#750)
* Set Ansible version in openstack CI for 2.3

Ansible 2.4 just came out and it breaks our playbooks. Let's pin the end
to end CI version to 2.3 until we've figured it out.

* Only use the deployed DNS for validation

I think the openshift installation and teardown is broken now that the
public DNS disables recursion. So we'll only use it for the validation
steps and than turn it back off.

* Use bash trap to clean up the DNS

* Actually display the commit message

* Use openshift-ansible-3.6.22-1 in openstack CI

The commit following this tag is broken for us.